Commit fast-build patch to portage_tool repository.
This patch allows us to selectively disable Portage locks in parallel_emerge.
It is not intended to be upstreamed but is left in for now until we find a way
to replace it (e.g. by integrating parallel emerge support directly into portage
itself).

Change-Id: I4ea50e206102b249ba01d2f760ebaba221bddd18

BUG=chromium-os:11324
TEST=build_packages --nousepkg

Review URL: http://codereview.chromium.org/6377009