re-order install hooks execution
We want the install hooks to run before stripping so we can do our own
QA checks on them.  Bump the order so that happens.

In the future, we'll want to re-architect this code entirely, but the
qa checking code in portage will need a rewrite.  See BUG for more
details.

BUG=chromium:217578
TEST=`emerge-x86-alex power_manager` shows flag stuff works and things still get stripped
TEST=`cbuildbot chromiumos-sdk` works

Change-Id: I00a940c1d4667554f92ec3bdb3786c12adeed79c
Reviewed-on: https://chromium-review.googlesource.com/172922
Reviewed-by: Marc MERLIN <merlin@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>