re-order install hooks execution
We want the install hooks to run before stripping so we can do our own
QA checks on them.  Bump the order so that happens.

In the future, we'll want to re-architect this code entirely, but the
qa checking code in portage will need a rewrite.  See BUG for more
details.

BUG=chromium:217578
TEST=`emerge-x86-alex power_manager` shows flag stuff works and things still get stripped
TEST=`cbuildbot chromiumos-sdk` works

https://chromium-review.googlesource.com/172922
https://chromium-review.googlesource.com/214527
https://chromium-review.googlesource.com/815659

Change-Id: Iebf90b4afba35a90a4259f220ac5bcf726ad6b4f
Reviewed-on: https://chromium-review.googlesource.com/c/1377093
Tested-by: Allen Webb <allenwebb@google.com>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Allen Webb <allenwebb@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/portage_tool/+/1865455
Reviewed-by: Allen Webb <allenwebb@google.com>
Commit-Queue: Chris McDonald <cjmcdonald@chromium.org>
Tested-by: Chris McDonald <cjmcdonald@chromium.org>