re-order install hooks execution
We want the install hooks to run before stripping so we can do our own
QA checks on them.  Bump the order so that happens.

In the future, we'll want to re-architect this code entirely, but the
qa checking code in portage will need a rewrite.  See BUG for more
details.

BUG=chromium:217578
TEST=`emerge-x86-alex power_manager` shows flag stuff works and things still get stripped
TEST=`cbuildbot chromiumos-sdk` works

https://chromium-review.googlesource.com/172922

Change-Id: I5b98037c594fc681b4dc0c9ebf6c0194fd9672b1
Reviewed-on: https://chromium-review.googlesource.com/214527
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: Bertrand Simonnet <bsimonnet@chromium.org>