Add separatedebug feature
When separatedebug is set in FEATURES, produce the debug symbols in a separate
tarball.
The resulting tarball is stored in PKGDIR with the extension .debug.tbz2.

BUG=chromium:432294
TEST=`FEATURES=separatedebug emerge-gizmo metrics` works.
/build/gizmo/packages/chromeos-base/metrics-0.0.1-r365.debug.tbz2 is created.
TEST=`emerge-gizmo metrics` works.
TEST=trybot run on daisy, link, gizmo, chromiumos-sdk, duck, lumpy-incremental.

https://chromium-review.googlesource.com/229173

binarytree: ignore separatedebug archives

When we use FEATURES=separatedebug, we create .debug.tbz2 files in the
binary tree that portage warns about.  Silence that here.

BUG=chromium:432294
TEST=using emerge doesn't issue "Binary package name is invalid" warnings

https://chromium-review.googlesource.com/932581

Change-Id: I4ba713e38987ec2800323ad8766b0d7cf1ee356e
Reviewed-on: https://chromium-review.googlesource.com/c/1377099
Tested-by: Allen Webb <allenwebb@google.com>
Commit-Queue: Allen Webb <allenwebb@google.com>
Reviewed-by: Mike Frysinger <vapier@chromium.org>