flat_hash: Fix race condition in _setitem.
When setting a value in a flat_hash cache, the pid is used to create a
unique temporary file. If two different processes with the same pid
(different pid namespaces) try to set the cache, the code is racy.

This change converts _setitem to use tempfile which does not rely on
the pid being unique.

Fix is part of v2.3.1 (5652bc88514bdb36b36b544f7fc7e623cf25caae).

BUG=chromium:477727
TEST=trybot run on chromiumos-sdk.

https://chromium-review.googlesource.com/270802

Change-Id: Idee01df326e0408778ad84d7721350c7d7c03ad8
Reviewed-on: https://chromium-review.googlesource.com/815664
Reviewed-by: Chirantan Ekbote <chirantan@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>