binarytree: clear build_time with binhosts
Before the multi-binhost support in commit 328dd4712f88cbb8ef390a,
portage would only consult one CPV.  This allowed us to control the
ordering based on the PORTAGE_BINHOST so we could have public full
builders (including a Chromium browser) and we could have internal
Chrome builds, and the last one would be preferred over the first.

After that commit, the build time is mixed in, and portage will
(reasonably) consider a newer build as better.  The trouble is that
our full builders run more often than our Chrome PFQ builders, so
the Chromium build will have a newer time, and we install that.

For now, lets clear the build_time setting with the binpkgs.  This
lets us retain the PORTAGE_BINHOST ordering until we can figure out
a better way of managing this, or upstream gets a knob for us.

We also don't bother inserting the new entry into the "metadata"
variable because it isn't used after this logic, and the only code
here that cares about it is the test for "is a local binpkg already
available".

https://bugs.gentoo.org/647076

BUG=chromium:809312
TEST=precq passes

https://chromium-review.googlesource.com/922807

Change-Id: I6eef52c6c92ab177de19ec51ca209f1bb86835b0
Reviewed-on: https://chromium-review.googlesource.com/c/1377100
Commit-Queue: Allen Webb <allenwebb@google.com>
Tested-by: Allen Webb <allenwebb@google.com>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/portage_tool/+/1865461
Reviewed-by: Allen Webb <allenwebb@google.com>
Commit-Queue: Chris McDonald <cjmcdonald@chromium.org>
Tested-by: Chris McDonald <cjmcdonald@chromium.org>