ebuild: hack the chost check for armv7a-cros-linux-gnueabihf
We are in the process of rename armv7a-cros-linux-gnueabi to
armv7a-cros-linux-gnueabihf. See crbug.com/711369
We created two sets to toolchains, namely armv7a-cros-linux-gnueabi
and armv7a-cros-linux-gnueabihf, they are basically then same with
different names.
When changing the CHOST for one board from armv7a-cros-linux-gnueabi
to armv7a-cros-linux-gnueabihf, it failed on pre-cq. Because in the
build_package stage, it calls function VerifyBinpkg to verify that
an binary package exists for chromeos-chrome. The check requires the
CHOST from board configure and prebuilt packages are the same, and the
change broke that.
This patch silence the error to make pre-cq pass.

BUG=chromium:711369
TEST=emerge-daisy -G zlib succeeded with new CHOST set.

Change-Id: I6ae231ebc663b2afc0c3773418258daa4f78d984
Reviewed-on: https://chromium-review.googlesource.com/1055211
Commit-Ready: Yunlian Jiang <yunlian@chromium.org>
Tested-by: Yunlian Jiang <yunlian@chromium.org>
Reviewed-by: Yunlian Jiang <yunlian@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>