Add support for a distinct SYSROOT location
This location cannot be prefix-aware as we don't currently have any
variable to represent that. Given that this is intended to be used
with crossdev toolchains under /usr/${CHOST}, that should not
matter. Even if that location is nested within a prefixed system, it
does not need to be aware of that as it is not intended for execution.

Cherry-picked from https://github.com/chewi/portage/tree/sysroot-fixes
Upstream status: Not sent

BUG=chromium:929503
TEST=build_packages

Change-Id: I98f95bc6f2ba9240174ee7d0602cae8868d3f4b1
Reviewed-on: https://chromium-review.googlesource.com/1686967
Tested-by: Chris McDonald <cjmcdonald@chromium.org>
Commit-Ready: Chris McDonald <cjmcdonald@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>