portage: Add OWNERS file
Enforcement is not enabled yet. This CL is a suggestion of ownership
candidates based on git activity and you are welcome to add other owners
as you see fit. See the following link for more information on the syntax
for OWNERS files in ChromiumOS repositories:
https://gerrit.googlesource.com/plugins/find-owners/+/master/src/main/resources/Documentation/syntax.md

BUG=chromium:748644
TEST=None

Change-Id: I007139e15806e67a08b4163bf5f6e42986d07e0d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/portage_tool/+/1597096
Tested-by: Chris McDonald <cjmcdonald@chromium.org>
Reviewed-by: Alex Klein <saklein@chromium.org>
Reviewed-by: Ned Nguyen <nednguyen@google.com>
Commit-Queue: Chris McDonald <cjmcdonald@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/portage_tool/+/1865465
Reviewed-by: Chris McDonald <cjmcdonald@chromium.org>