elog: fix log flushing
When output is being piped to a file, the use of print (but no flush)
intermingled with EOutput logs (which can go to stderr & flush) can
result in the output being coalesced incorrectly: we get all of the
messages in stderr flushed first, then all the blank lines.

Add an explicit flush after each print so they stay in sync with the
flushed output from the EOutput logic.

BUG=chromium:996480
TEST=`sudo emerge libaio |& tee /dev/null` prints correctly

Change-Id: Iac21d9900460f29c1a69972dc0f2b1f379358d7e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/portage_tool/+/1961314
Reviewed-by: Chris McDonald <cjmcdonald@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>