CHROMIUM: Log completion of package installs
The log currently shows Emerging/Installing progress messages, but
it doesn't indicate when a package finishes.  Since there can be dozens
of installs in flight at once, this makes it nearly impossible to
determine how long an individual package took.  This adds a similar
"Completed" line when the package merge completes.

BUG=None
TEST=sudo emerge portage; build_packages

Change-Id: Ida0b2aedc2a12a910362fe32e1f04955c7dad50b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/portage_tool/+/2037661
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Chris McDonald <cjmcdonald@chromium.org>
Tested-by: Benjamin Gordon <bmgordon@chromium.org>