CHROMIUM: src_configure: dump config.log on failures
When configure fails, it is often with short terse messages, and
people need to dive into the config.log to find out why.  On bots,
this is not possible as the logs are discarded.  Dump it to stdout
for people to have immediate access to.

This is what we want on bots, but will be very verbose for local
development.  Not sure if we can or should improve on this.  We
can start with this and see how people react.

BUG=None
TEST=CQ passes

Change-Id: I68f8ad117ea7449b31620d5e144f6d0b7a8e803e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/portage_tool/+/2223827
Reviewed-by: Chris McDonald <cjmcdonald@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>